+++ src/Kernel/CryptographyHelper.h
@@ -8,6 +8,6 @@ namespace Mengine
     namespace Helper
     {
         bool generateRandomSeed( uint64_t * const _seed );
-        bool generateRandomHexadecimal( uint32_t _length, Char * const _hexadecimal, bool _lowercase );
+        bool generateRandomHexadecimal( size_t _length, Char * const _hexadecimal, bool _lowercase );
     }
 }
